Modifications of strategies for the reproducible experiment diagram

+++ jf-engine-instance/src/main/java/fr/axonic/jf/instance/ReproducibleExperiment/evidences/AccuracyReproducibilityThreshold.java
@@ -7,7 +7,7 @@ import javax.xml.bind.annotation.XmlRootElement;
 @XmlRootElement
 public class AccuracyReproducibilityThreshold extends Evidence<ThresholdDocument> {
     public static final String EVIDENCE_NAME = "ACCURACY_THRESHOLD_EVIDENCE";
-    public static final int VALUE = 0;
+    public static final double VALUE = 0.0;
 
     public AccuracyReproducibilityThreshold(ThresholdDocument element) {
         super(EVIDENCE_NAME, element);
