+++ jf-engine-instance/src/main/java/fr/axonic/jf/instance/ReproducibleExperiment/evidences/TotalTimeReproducibilityThreshold.java
@@ -8,7 +8,7 @@ import javax.xml.bind.annotation.XmlRootElement;
 @XmlRootElement
 public class TotalTimeReproducibilityThreshold extends Evidence<ThresholdDocument> {
     public static final String EVIDENCE_NAME = "TOTALTIME_THRESHOLD_EVIDENCE";
-    public static final int VALUE = 0;
+    public static final double VALUE = 0.0;
 
     public TotalTimeReproducibilityThreshold(ThresholdDocument element) {
         super(EVIDENCE_NAME, element);
